+++ nucleotides/task/short_read_assembler.py
@@ -1,12 +1,15 @@
-import os.path, shutil
+import os.path, shutil, funcy
 
-import biobox_cli.util.misc          as bbx_util
-import nucleotides.util              as util
-import nucleotides.command.run_image as image
-import nucleotides.metrics           as met
+import nucleotides.metrics     as met
+import nucleotides.filesystem  as fs
 
-def successful_event_outputs():
-    return set(["contig_fasta"])
+def biobox_args(app):
+    path = fs.get_input_file_path('short_read_fastq', app)
+    return [{"fastq" : [
+        {"id" : 0 , "value" : path, "type": "paired"}]}]
+
+def output_files():
+    return [['contig_fasta', [0, 'fasta', 0]]]
 
 def collect_metrics(app):
     import json
@@ -17,17 +20,5 @@ def collect_metrics(app):
     else:
         return {}
 
-def setup(app):
-    None
-
-def create_biobox_args(app):
-    return ["run",
-            app["task"]["image"]["type"],
-            app["task"]["image"]["name"],
-            "--input={}".format(image.get_input_file_path('short_read_fastq', app)),
-            "--output={}".format(image.get_tmp_file_path('contig_fasta', app)),
-            "--task={}".format(app["task"]["image"]["task"]),
-            "--no-rm"]
-
-def copy_output_files(app):
-    image.copy_tmp_file_to_outputs(app, 'contig_fasta', 'contig_fasta')
+def successful_event_outputs():
+    return set(["contig_fasta"])