+++ test/helper/application.py
@@ -1,6 +1,7 @@
-import os
+import os, shutil, json
+
+import boltons.fileutils     as fu
 import helper.file           as file_helper
-import biobox_cli.util.misc  as bbx_util
 import nucleotides.log       as log
 
 def mock_app():
@@ -10,8 +11,18 @@ def mock_app():
            'path'   : path}
     return app
 
-def mock_short_read_assembler_state(task = True, dummy_reads = False, reads = False, outputs = False):
-    import json, shutil
+def copy_to_directory(src_file, dst_dir, app):
+    dir_ = os.path.join(app['path'], dst_dir)
+    fu.mkdir_p(dir_)
+    shutil.copy(src_file, dir_)
+
+def copy_to_file(src_file, dst_file, app):
+    file_ = os.path.join(app['path'], dst_file)
+    dir_  = os.path.dirname(file_)
+    fu.mkdir_p(dir_)
+    shutil.copy(src_file, file_)
+
+def mock_short_read_assembler_state(task = True, dummy_reads = False, reads = False, intermediates = False, outputs = False):
     app = mock_app()
 
     if task:
@@ -20,22 +31,22 @@ def mock_short_read_assembler_state(task = True, dummy_reads = False, reads = Fa
             app["task"] = json.loads(f.read())
 
     if dummy_reads:
-        bbx_util.mkdir_p(app['path'] + '/inputs/short_read_fastq/')
-        shutil.copy('tmp/data/dummy.reads.fq.gz', app['path'] + '/inputs/short_read_fastq/')
+        copy_to_directory('tmp/data/dummy.reads.fq.gz', 'inputs/short_read_fastq', app)
 
     if reads:
-        bbx_util.mkdir_p(app['path'] + '/inputs/short_read_fastq/')
-        shutil.copy('tmp/data/reads.fq.gz', app['path'] + '/inputs/short_read_fastq/')
+        copy_to_directory('tmp/data/reads.fq.gz', 'inputs/short_read_fastq', app)
+
+    if intermediates:
+        copy_to_file('data/sra_biobox.yaml', 'tmp/biobox.yaml', app)
+        copy_to_directory('tmp/data/contigs.fa', 'tmp', app)
 
     if outputs:
-        bbx_util.mkdir_p(app['path'] + '/outputs/contig_fasta/')
-        shutil.copy('tmp/data/contigs.fa', app['path'] + '/outputs/contig_fasta/contigs')
-        bbx_util.mkdir_p(app['path'] + '/outputs/container_runtime_metrics/')
-        shutil.copy('tmp/data/container_runtime.json', app['path'] + '/outputs/container_runtime_metrics/metrics.json')
+        copy_to_directory('tmp/data/contigs.fa', 'outputs/contig_fasta', app)
+        copy_to_directory('tmp/data/metrics.json', 'outputs/container_runtime_metrics', app)
 
     return app
 
-def mock_reference_evaluator_state(outputs = False):
+def mock_reference_evaluator_state(intermediates = False, outputs = False):
     import json, shutil
     app = mock_app()
 
@@ -45,16 +56,15 @@ def mock_reference_evaluator_state(outputs = False):
 
     app["s3-upload"] = "s3://"
 
-    bbx_util.mkdir_p(app['path'] + '/inputs/reference_fasta/')
-    shutil.copy('tmp/data/reference.fa', app['path'] + '/inputs/reference_fasta/6bac51cc35.fa')
+    copy_to_file('tmp/data/reference.fa', 'inputs/reference_fasta/6bac51cc35.fa', app)
+    copy_to_file('tmp/data/contigs.fa', 'inputs/contig_fasta/7e9f760161.fa', app)
 
-    bbx_util.mkdir_p(app['path'] + '/inputs/contig_fasta/')
-    shutil.copy('tmp/data/contigs.fa', app['path'] + '/inputs/contig_fasta/7e9f760161.fa')
+    if intermediates:
+        copy_to_file('tmp/data/assembly_metrics.tsv', 'tmp/combined_quast_output/report.tsv', app)
+        copy_to_file('data/quast_biobox.yaml', 'tmp/biobox.yaml', app)
 
     if outputs:
-        bbx_util.mkdir_p(app['path'] + '/outputs/assembly_metrics/')
-        shutil.copy('tmp/data/assembly_metrics.tsv', app['path'] + '/outputs/assembly_metrics/outputs.csv')
-        bbx_util.mkdir_p(app['path'] + '/outputs/container_runtime_metrics/')
-        shutil.copy('tmp/data/container_runtime.json', app['path'] + '/outputs/container_runtime_metrics/metrics.json')
+        copy_to_file('tmp/data/assembly_metrics.tsv', 'outputs/assembly_metrics/outputs.csv', app)
+        copy_to_directory('tmp/data/metrics.json', 'outputs/container_runtime_metrics', app)
 
     return app