+++ test/helper/file.py
@@ -1,17 +1,17 @@
 import os, tempfile
-import biobox_cli.util.misc as bbx_util
-import nose.tools as nose
+import boltons.fileutils  as fu
+import nose.tools         as nose
 
 def create_benchmark_file(app, path, contents = ''):
     loc = os.path.join(app["path"] + path)
-    bbx_util.mkdir_p(os.path.dirname(loc))
+    fu.mkdir_p(os.path.dirname(loc))
     with open(loc, 'w+') as f:
         f.write(contents)
     return loc
 
 def test_dir():
     path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', 'tmp', 'tests')
-    bbx_util.mkdir_p(path)
+    fu.mkdir_p(path)
     return tempfile.mkdtemp(dir = path)
 
 def assert_is_dir(path):