+++ test/test_filesystem.py
@@ -1,0 +1,8 @@
+import nose.tools             as nose
+import helper.application     as app_helper
+import nucleotides.filesystem as fs
+
+def test_get_output_biobox_file_arguments():
+    app = app_helper.mock_short_read_assembler_state(intermediates = True)
+    args = fs.get_output_biobox_file_arguments(app)
+    nose.assert_in('fasta', args[0])